<Full Notes on BarScanner Implementation, and retreiveKegData Implemented>
-Full Notes of Last Commit(omitted during previous commit due to time constraints)-
Commented out code that used the old ACR scanner and replaced it with code to handle the new marson NFC scanner(refered to in code as the Bar Scanner). Added code to take in an id and request all the data about that keg's current/last known location from the servlet and display it as a sentance on the app.